requirements.txt
gradio
transformers
torch
openai-whisper
python-docx
pandas
